fix: address review findings on #1830 — move I/O-touching synthesizeSpeech dispatch tests from voice.unit.test.ts to voice.integration.test.ts
