+++ agent/src/voice.integration.test.ts
@@ -10,7 +10,7 @@
 import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
 import type { ChildProcess, SpawnOptions } from "node:child_process";
 import { EventEmitter } from "node:events";
-import { existsSync, unlinkSync, writeFileSync } from "node:fs";
+import { existsSync, statSync, unlinkSync, writeFileSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import type { VoiceConfig, WhisperClientFn } from "./voice.ts";
@@ -510,5 +510,157 @@ describe("transcribeAudio — fallback chain", () => {
 
     console.error = origError;
     expect(result).toBeNull();
+  });
+});
+
+// ─── Helpers for subprocess injection (Piper dispatch) ───────────────────────
+
+// Builds a fake ChildProcess and schedules `onImmediate` to fire on it once
+// stdout/stderr/stdin are wired up — mirrors the shape spawnFn callers expect
+// (proc.stdout/stderr as EventEmitters, proc.stdin.write/end) without a real
+// subprocess. `onImmediate` also receives the args the real spawnPiper call
+// was made with, so callers can locate the `--output_file` path a real
+// Piper binary would write to.
+function makeFakePiperSpawn(
+  onImmediate: (proc: EventEmitter, args: string[]) => void,
+): SpawnFn {
+  return (_cmd, args, _opts) => {
+    // biome-ignore lint/suspicious/noExplicitAny: test helper needs dynamic property assignment
+    const proc: any = new EventEmitter();
+    proc.stdout = new EventEmitter();
+    proc.stderr = new EventEmitter();
+    proc.stdin = { write: () => {}, end: () => {} };
+    setImmediate(() => onImmediate(proc, args));
+    return proc as unknown as ChildProcess;
+  };
+}
+
+// Extracts the path following `--output_file` from the args synthesizePiper
+// spawns with, mirroring how the real piper binary locates its output path.
+function outputFileFromArgs(args: string[]): string {
+  const idx = args.indexOf("--output_file");
+  if (idx === -1 || idx === args.length - 1) {
+    throw new Error("--output_file not found in spawn args");
+  }
+  return args[idx + 1];
+}
+
+// Mirrors what the real piper binary does: writes non-empty audio bytes to
+// the `--output_file` path before exiting 0. Lets tests assert a real file
+// was written, not just that a path string was returned.
+const makeSuccessSpawnPiper = () =>
+  makeFakePiperSpawn((proc, args) => {
+    writeFileSync(outputFileFromArgs(args), Buffer.from([1, 2, 3, 4]));
+    proc.emit("close", 0);
+  });
+
+const makeFailExitSpawnPiper = (code: number) =>
+  makeFakePiperSpawn((proc) => {
+    // biome-ignore lint/suspicious/noExplicitAny: fake proc's stderr is untyped
+    (proc as any).stderr.emit("data", Buffer.from("piper: model load error"));
+    proc.emit("close", code);
+  });
+
+const makeSpawnErrorSpawnPiper = () =>
+  makeFakePiperSpawn((proc) =>
+    proc.emit("error", new Error("ENOENT: piper binary not found")),
+  );
+
+// ─── synthesizeSpeech dispatch — ElevenLabs vs Piper ─────────────────────────
+
+describe("synthesizeSpeech — dispatch (ElevenLabs vs Piper)", () => {
+  // Spy on console.error by swapping the bound method (not a global
+  // override) and restoring it after each test — mirrors the pattern in
+  // piper-voice.unit.test.ts.
+  let errorCalls: unknown[][] = [];
+  let originalError: typeof console.error;
+
+  beforeEach(() => {
+    errorCalls = [];
+    originalError = console.error;
+    console.error = (...args: unknown[]) => {
+      errorCalls.push(args);
+    };
+  });
+
+  afterEach(() => {
+    console.error = originalError;
+  });
+
+  test("with ELEVENLABS_API_KEY set, ElevenLabs is called and Piper's spawnFn is never invoked", async () => {
+    const mockFetch = mock(
+      async () => new Response(new Uint8Array([1, 2, 3]).buffer, { status: 200 }),
+    );
+    const spawnFn = mock((() => {
+      throw new Error("spawnFn should not be called when ElevenLabs key is set");
+    }) as unknown as SpawnFn);
+
+    const outPath = await synthesizeSpeech(
+      "hello",
+      { elevenLabsApiKey: "test-eleven-key" },
+      mockFetch as unknown as typeof fetch,
+      spawnFn,
+    );
+
+    expect(mockFetch).toHaveBeenCalledTimes(1);
+    expect(spawnFn).not.toHaveBeenCalled();
+    expect(outPath).not.toBeNull();
+    expect(outPath).toContain(".mp3");
+    if (outPath) unlinkSync(outPath);
+  });
+
+  test("with no ELEVENLABS_API_KEY and a working Piper, synthesizeSpeech returns a path to a non-empty .wav file", async () => {
+    const outPath = await synthesizeSpeech(
+      "hello",
+      {},
+      undefined,
+      makeSuccessSpawnPiper(),
+    );
+
+    expect(outPath).not.toBeNull();
+    expect(outPath).toContain(".wav");
+    if (outPath) {
+      expect(existsSync(outPath)).toBe(true);
+      expect(statSync(outPath).size).toBeGreaterThan(0);
+      unlinkSync(outPath);
+    }
+  });
+
+  test("when Piper exits non-zero, synthesizeSpeech returns null with no throw", async () => {
+    let result: string | null = null;
+    let threw = false;
+    try {
+      result = await synthesizeSpeech(
+        "hello",
+        {},
+        undefined,
+        makeFailExitSpawnPiper(1),
+      );
+    } catch {
+      threw = true;
+    }
+
+    expect(threw).toBe(false);
+    expect(result).toBeNull();
+    expect(errorCalls.length).toBeGreaterThan(0);
+  });
+
+  test("when the Piper spawn itself errors (binary absent), synthesizeSpeech returns null with no throw", async () => {
+    let result: string | null = null;
+    let threw = false;
+    try {
+      result = await synthesizeSpeech(
+        "hello",
+        {},
+        undefined,
+        makeSpawnErrorSpawnPiper(),
+      );
+    } catch {
+      threw = true;
+    }
+
+    expect(threw).toBe(false);
+    expect(result).toBeNull();
+    expect(errorCalls.length).toBeGreaterThan(0);
   });
 });
